Fixed some stuff
Firstly, the "clear the map" button now works.

Secondly, the map will now centre between the start and end coords when pressing "Find path" or "Clear the map".

Thirdly, there will be error msg pop up when the source or destination textboxes have non-float OR != 2 coords in each box.

Fourthly, the left side main_layout elements now scale to 1/4 of the left side of the screen while the right side webView will scale to the other 3/4 of the screen no matter if u fullscreen or resize (apart being unable to minize below the setMinimumSize value)